requirements.txt
streamlit
azure-storage-blob
PyPDF2
base64